+++ README.md
@@ -1,11 +1,8 @@
 # DiaryApi
 
-[![Coverage Status](https://coveralls.io/repos/github/ibraheemkabir/DiaryApi/badge.svg?branch=master)](https://coveralls.io/github/ibraheemkabir/DiaryApi?branch=master)
-
 [![Build Status](https://travis-ci.org/ibraheemkabir/DiaryApi.svg?branch=master)](https://travis-ci.org/ibraheemkabir/DiaryApi)
 
-[![Test Coverage](https://api.codeclimate.com/v1/badges/26e3517bc3052f064239/test_coverage)](https://codeclimate.com/github/ibraheemkabir/DiaryApi/test_coverage)
-
+[![Coverage Status](https://coveralls.io/repos/github/ibraheemkabir/DiaryApi/badge.svg?branch=ft-getspecificentry-%23159179157)](https://coveralls.io/github/ibraheemkabir/DiaryApi?branch=ft-getspecificentry-%23159179157) 
 
 [![Maintainability](https://api.codeclimate.com/v1/badges/26e3517bc3052f064239/maintainability)](https://codeclimate.com/github/ibraheemkabir/DiaryApi/maintainability)
