+++ app.js
@@ -1,17 +1,21 @@
-import express from 'express';
-import bodyParser from 'body-parser';
-
-const routes = require('./routes/deleteEntry');
+const bodyParser = require('body-parser');
+const express = require('express');
 
 const app = express();
 
-app.use(express.json());
-app.use(bodyParser.json());
+
 app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
 
-app.use('/api/v1/entry', routes);
 
-const port = process.env.Port || 3000;
-app.listen(port, () => console.log(`Listening on port ${port}...`));
+const del = require('./server/routes/deleteEntry');
+
+
+app.use('/api/v1/entry/', del);
+
+app.set('port', (process.env.PORT || 3000));
+app.listen(app.get('port'), () => {
+  console.log('Listening on port...');
+});
 
 module.exports = app;
