+++ package.json
@@ -4,7 +4,6 @@
   "description": "",
   "main": "Api.js",
   "dependencies": {
-    "babel": "^6.23.0",
     "body-parser": "^1.18.3",
     "chai": "^4.1.2",
     "chai-http": "^4.0.0",
@@ -12,13 +11,15 @@
     "express-jwt": "^5.3.1",
     "jsonwebtoken": "^8.3.0",
     "node-datetime": "^2.0.6",
+    "nodemon": "^1.18.3",
     "travis": "^0.1.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-latest": "^6.24.1",
     "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "chai-http": "^4.0.0",
     "coveralls": "^3.0.2",
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^17.0.0",
@@ -27,15 +28,13 @@
     "eslint-plugin-react": "^7.10.0",
     "istanbul": "^0.4.5",
     "mocha": "^5.2.0",
-    "nodemon": "^1.18.3"
+    "nyc": "^12.0.2"
   },
   "scripts": {
-    "babel-node": "babel-node --presets=es2015",
-    "start": "nodemon --exec npm run babel-node -- ./app.js",
+    "start": "./node_modules/.bin/babel-node app.js",
     "build": "babel src -d dist",
-    "coverage": "node node_modules/.bin/istanbul cover _mocha -- -R spec",
-    "test": "mocha --compilers js:babel-core/register --exit",
-    "coveralls": "cat ./coverage/lcov.info | node node_modules/.bin/coveralls"
+    "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "test": " nyc mocha --compilers js:babel-core/register server/test/ --exit"
   },
   "keywords": [],
   "author": "",
