+++ server/helpers/array.js
@@ -1,0 +1,9 @@
+import User from '../models/entrymodel';
+
+const id = 1;
+
+const entries = [
+  new User(id, 'Dear diary today is my fist post', 'my first post', '29-34-2016'),
+];
+
+export default entries;
