+++ server/models/entrymodel.js
@@ -1,0 +1,9 @@
+
+export default class User {
+  constructor(id, title, content, date) {
+    this.id = id;
+    this.title = title;
+    this.content = content;
+    this.date = date;
+  }
+}
